Documentation

Documentation

ChemometricsTools.BlandAltman — Method.

BlandAltman(Y1, Y2; Confidence = 1.96)

Returns a Plot object of a Bland-Altman plot between vectors Y1 and Y2 with a confidence limit of Confidence.

source

ChemometricsTools.Bounds — Method.

Bounds(dims)

Constructor for a Bounds object. Returns a bounds object with a lower bound of [lower...] and upper bound[upper...] with length of dims.

source

ChemometricsTools.Bounds — Method.

Bounds(dims)

Default constructor for a Bounds object. Returns a bounds object with a lower bound of [0...] and upper bound[1...] with length of dims.

source

ChemometricsTools.CORAL — Method.

CORAL(X1, X2; lambda = 1.0)

Performs CORAL to facilitate covariance based transfer from X1 to X2 with regularization parameter lambda. Returns a CORAL object.

Correlation Alignment for Unsupervised Domain Adaptation. Baochen Sun, Jiashi Feng, Kate Saenko. https://arxiv.org/abs/1612.01939

source

ChemometricsTools.CORAL — Method.

(C::CORAL)(Z)

Applies a the transform from a learned CORAL object to new data Z.

source

ChemometricsTools.KFoldsValidation — Method.

KFoldsValidation(K::Int, x, y)

Returns a KFoldsValidation iterator with K folds. Because it's an iterator it can be used in for loops, see the tutorials for pragmatic examples. The iterator returns a 2-Tuple of 2-Tuples which have the following form: ((TrainX,TrainY),(ValidateX,ValidateY).

source

ChemometricsTools.Kernel — Method.

(K::Kernel)(X)

This is a convenience function to allow for one-line construction of kernels from a Kernel object K and new data X.

source

ChemometricsTools.LDA — Method.

LDA(X, Y; Factors = 1)

Compute's a LinearDiscriminantAnalysis transform from x with a user specified number of latent variables(Factors). Returns an LDA object.

source

ChemometricsTools.LDA — Method.

( model::LDA )( Z; Factors = length(model.Values) )

Calling a LDA object on new data brings the new data Z into the LDA basis.

source

ChemometricsTools.MultiplicativeScatterCorrection — Method.

(T::MultiplicativeScatterCorrection)(Z)

Applies MultiplicativeScatterCorrection from a stored object T to Array Z.

source

ChemometricsTools.OrthogonalSignalCorrection — Method.

OrthogonalSignalCorrection(X, Y; Factors = 1)

Performs Thomas Fearn's Orthogonal Signal Correction to an endogenous X and exogenous Y. The number of Factors are the number of orthogonal components to be removed from X. This function returns an OSC object.

Tom Fearn. On orthogonal signal correction. Chemometrics and Intelligent Laboratory Systems. Volume 50, Issue 1, 2000, Pages 47-52.

source

ChemometricsTools.OrthogonalSignalCorrection — Method.

(OSC::OrthogonalSignalCorrection)(Z; Factors = 2)

Applies a the transform from a learned orthogonal signal correction object OSC to new data Z.

source

ChemometricsTools.PCA — Method.

PCA(X; Factors = minimum(size(X)) - 1)

Compute's a PCA from x using LinearAlgebra's SVD algorithm with a user specified number of latent variables(Factors). Returns a PCA object.

source

ChemometricsTools.PCA — Method.

(T::PCA)(Z::Array; Factors = length(T.Values), inverse = false)

Calling a PCA object on new data brings the new data Z into or out of (inverse = true) the PCA basis.

source

ChemometricsTools.Particle — Method.

Particle(ProblemBounds, VelocityBounds)

Default constructor for a Particle object. It creates a random unformly distributed particle within the specified ProblemBounds, and limits it's velocity to the specified VelocityBounds.

source

ChemometricsTools.QQ — Method.

QQ( Y1, Y2; Quantiles = collect( 1 : 99 ) ./ 100 )

Returns a Plot object of a Quantile-Quantile plot between vectors Y1 and Y2 at the desired Quantiles.

source

ChemometricsTools.RandomForest — Type.

RandomForest(x, y, mode = :classification; gainfn = entropy, trees = 50, maxdepth = 10,  minbranchsize = 5, samples = 0.7, maxvars = nothing)

Returns a classification (mode = :classification) or a regression (mode = :regression) random forest model. The gainfn can be entropy or gini for classification or ssd for regression. If the number of maximumvars is not provided it will default to sqrt(variables) for classification or variables/3 for regression.

The returned object can be used for inference by calling new data on the object as a function.

Breiman, L. Machine Learning (2001) 45: 5. https://doi.org/10.1023/A:1010933404324

source

ChemometricsTools.RandomForest — Method.

(RF::RandomForest)(X)

Returns bagged prediction vector of random forest model.

source

ChemometricsTools.RollingWindow — Method.

RollingWindow(samples::Int,windowsize::Int,skip::Int)

Creates a RollingWindow iterator from a number of samples and a static windowsize where every iteration skip steps are skipped. The iterator can be used in for loops to iteratively return indices of a dynamic rolling window.

source

ChemometricsTools.RollingWindow — Method.

RollingWindow(samples::Int,windowsize::Int)

Creates a RollingWindow iterator from a number of samples and a static windowsize. The iterator can be used in for loops to iteratively return indices of a dynamic rolling window.

source

ChemometricsTools.RunningMean — Method.

RunningMean(x)

Constructs a running mean object with an initial scalar value of x.

source

ChemometricsTools.RunningVar — Method.

RunningVar(x)

Constructs a RunningVar object with an initial scalar value of x. Note: RunningVar objects implicitly calculate the running mean.

source

ChemometricsTools.TransferByOrthogonalProjection — Method.

TransferByOrthogonalProjection(X1, X2; Factors = 1)

Performs Thomas Fearns Transfer By Orthogonal Projection to facilitate transfer from X1 to X2. Returns a TransferByOrthogonalProjection object.

Anne Andrew, Tom Fearn. Transfer by orthogonal projection: making near-infrared calibrations robust to between-instrument variation. Chemometrics and Intelligent Laboratory Systems. Volume 72, Issue 1, 2004, Pages 51-56,

source

ChemometricsTools.TransferByOrthogonalProjection — Method.

(TbOP::TransferByOrthogonalProjection)(X1; Factors = TbOP.Factors)

Applies a the transform from a learned transfer by orthogonal projection object TbOP to new data X1.

source

ChemometricsTools.ALSSmoother — Method.

ALSSmoother(y; lambda = 100, p = 0.001, maxiters = 10)

Applies an assymetric least squares smoothing function to a vector y. The lambda, p, and maxiters parameters control the smoothness. See the reference below for more information.

Paul H. C. Eilers, Hans F.M. Boelens. Baseline Correction with Asymmetric Least Squares Smoothing. 2005

source

ChemometricsTools.BTEM — Function.

BTEM(X, bands = nothing; Factors = 3, particles = 50, maxiters = 1000)

Returns a single recovered spectra from a 2-Array X, the selected bands, number of Factors, using a Particle Swarm Optimizer.

Note: This is not the function used in the original paper. This will be updated... it was written from memory. Also the original method uses Simulated Annealing not PSO. Band-Target Entropy Minimization (BTEM): An Advanced Method for Recovering Unknown Pure Component Spectra. Application to the FTIR Spectra of Unstable Organometallic Mixtures. Wee Chew,Effendi Widjaja, and, and Marc Garland. Organometallics 2002 21 (9), 1982-1990. DOI: 10.1021/om0108752

source

ChemometricsTools.BTEMobjective — Method.

BTEMobjective( a, X )

Returns the scalar BTEM objective function obtained from the linear combination vector a and loadings X.

Note: This is not the function used in the original paper. This will be updated... it was written from memory.

source

ChemometricsTools.ClassificationTree — Method.

ClassificationTree(x, y; gainfn = entropy, maxdepth = 4, minbranchsize = 3)

Builds a CART object using either gini or entropy as a partioning method. Y must be a one hot encoded 2-Array. Predictions can be formed by calling the following function from the CART object: (M::CART)(x).

*Note: this is a purely nonrecursive decision tree. The julia compiler doesn't like storing structs of nested things. I wrote it the recursive way in the past and it was quite slow, I think this is true also of interpretted languages like R/Python...So here it is, nonrecursive tree's!

source

ChemometricsTools.DirectStandardization — Method.

DirectStandardization(InstrumentX1, InstrumentX2; Factors = minimum(collect(size(InstrumentX1))) - 1)

Makes a DirectStandardization object to facilitate the transfer from Instrument #2 to Instrument #1 . The returned object can be used to transfer unseen data to the approximated space of instrument 1. The number of Factors used are those from the internal orthogonal basis.

Yongdong Wang and Bruce R. Kowalski, "Calibration Transfer and Measurement Stability of Near-Infrared Spectrometers," Appl. Spectrosc. 46, 764-771 (1992)

source

ChemometricsTools.EWMA — Method.

EWMA(Initial::Float64, Lambda::Float64) = ewma(Lambda, Initial, Initial, RunningVar(Initial))

Constructs an exponentially weighted moving average object from an vector of scalar property values Initial and the decay parameter Lambda. This computes the running statistcs neccesary for creating the EWMA model using the interval provided and updates the center value to the mean of the provided values.

source

ChemometricsTools.EWMA — Method.

EWMA(Initial::Float64, Lambda::Float64) = ewma(Lambda, Initial, Initial, RunningVar(Initial))

Constructs an exponentially weighted moving average object from an initial scalar property value Initial and the decay parameter Lambda. This defaults the center value to be the initial value.

source

ChemometricsTools.EmpiricalQuantiles — Method.

EmpiricalQuantiles(X, quantiles)

Finds the column-wise quantiles of 2-Array X and returns them in a 2-Array of size quantiles by variables. *Note: This copies the array... Use a subset if memory is the concern. *

source

ChemometricsTools.EuclideanDistance — Method.

EuclideanDistance(X, Y)

Returns the euclidean distance matrix of X and Y such that the columns are the samples in Y.

source

ChemometricsTools.EuclideanDistance — Method.

EuclideanDistance(X)

Returns the Grahm aka the euclidean distance matrix of X.

source

ChemometricsTools.ExplainedVariance — Method.

ExplainedVariance(lda::LDA)

Calculates the explained variance of each singular value in an LDA object.

source

ChemometricsTools.ExplainedVariance — Method.

ExplainedVariance(PCA::PCA)

Calculates the explained variance of each singular value in a pca object.

source

ChemometricsTools.FNNLS — Method.

FNNLS(A, b; LHS = false, maxiters = 520)

Uses an implementation of Bro et. al's Fast Non-Negative Least Squares on the matrix A and vector b. We can state whether to pose the problem has a left-hand side problem (LHS = true) or a right hand side problem (default). Returns regression coefficients in the form of a vector.

Note: this function does not have guarantees. Use at your own risk for now. Fast Non-Negative Least Squares algorithm based on Bro, R., & de Jong, S. (1997) A fast non-negativity-constrained least squares algorithm. Journal of Chemometrics, 11, 393-401.

source

ChemometricsTools.FirstDerivative — Method.

FirstDerivative(X)

Uses the finite difference method to compute the first derivative for every row in X. Note: This operation results in the loss of a column dimension.

source

ChemometricsTools.FractionalDerivative — Function.

FractionalDerivative(Y, X = 1 : length(Y); Order = 0.5)

Calculates the Grunwald-Leitnikov fractional order derivative on every row of Array Y. Array X is a vector that has the spacing between column-wise entries in Y. X can be a scalar if that is constant (common in spectroscopy). Order is the fractional order of the derivative. Note: This operation results in the loss of a column dimension.

The Fractional Calculus, by Oldham, K.; and Spanier, J. Hardcover: 234 pages. Publisher: Academic Press, 1974. ISBN 0-12-525550-0

source

ChemometricsTools.GaussianKernel — Method.

GaussianKernel(X, Y, sigma)

Creates a Gaussian/RBF kernel from Arrays X and Y with hyperparameter sigma.

source

ChemometricsTools.GaussianKernel — Method.

GaussianKernel(X, sigma)

Creates a Gaussian/RBF kernel from Array X using hyperparameter sigma.

source

ChemometricsTools.Hotelling — Method.

Hotelling(X, pca::PCA; Quantile = 0.05, Variance = 1.0)

Computes the hotelling Tsq and upper control limit cut off of a pca object using a specified Quantile and cumulative variance explained Variance for new or old data X.

A review of PCA-based statistical process monitoring methodsfor time-dependent, high-dimensional data. Bart De Ketelaere https://wis.kuleuven.be/stat/robust/papers/2013/deketelaere-review.pdf

source

ChemometricsTools.IntervalOverlay — Method.

IntervalOverlay(Spectra, Intervals, Err)

Displays the relative error(Err) of each interval ontop of a Spectra.

source

ChemometricsTools.KennardStone — Method.

KennardStone(X, TrainSamples; distance = "euclidean")

Returns the indices of the Kennard-Stone sampled exemplars (E), and those not sampled (O) as a 2-Tuple (E, O).

R. W. Kennard & L. A. Stone (1969) Computer Aided Design of Experiments, Technometrics, 111, 137-148, DOI: 10.1080/00401706.1969.10490666

source

ChemometricsTools.Leverage — Method.

Leverage(pca::PCA)

Calculates the leverage of samples in a pca object.

source

ChemometricsTools.Limits — Method.

Limits(P::ewma; k = 3.0)

This function returns the upper and lower control limits with a k span of variance for an EWMA object P.

source

ChemometricsTools.LinearKernel — Method.

LinearKernel(X, Y, c)

Creates a Linear kernel from Arrays X and Y with hyperparameter C.

source

ChemometricsTools.LinearKernel — Method.

LinearKernel(X, c)

Creates a Linear kernel from Array X and hyperparameter C.

source

ChemometricsTools.MAE — Method.

MAE( y, yhat )

Calculates Mean Average Error from vectors Y and YHat

source

ChemometricsTools.MAPE — Method.

MAPE( y, yhat )

Calculates Mean Average Percent Error from vectors Y and YHat

source

ChemometricsTools.MCRALS — Function.

MCRALS(X, C, S = nothing; norm = (false, false), Factors = 1, maxiters = 20, nonnegative = (false, false) )

Performs Multivariate Curve Resolution using Alternating Least Squares on X taking initial estimates for S or C. S or C can be constrained by their norm, or by nonnegativity using nonnegative arguments. The number of resolved Factors can also be set.

Tauler, R. Izquierdo-Ridorsa, A. Casassas, E. Simultaneous analysis of several spectroscopic titrations with self-modelling curve resolution.Chemometrics and Intelligent Laboratory Systems. 18, 3, (1993), 293-300.

source

ChemometricsTools.ME — Method.

ME( y, yhat )

Calculates Mean Error from vectors Y and YHat

source

ChemometricsTools.MSE — Method.

MSE( y, yhat )

Calculates Mean Squared Error from vectors Y and YHat

source

ChemometricsTools.MakeIntervals — Function.

MakeIntervals( columns::Int, intervalsize::Union{Array, Tuple} = [20, 50, 100] )

Creates an Dictionary whose key is the interval size and values are an array of intervals from the range: 1 - columns of size intervalsize.

source

ChemometricsTools.MakeIntervals — Function.

MakeIntervals( columns::Int, intervalsize::Int = 20 )

Returns an 1-Array of intervals from the range: 1 - columns of size intervalsize.

source

ChemometricsTools.ManhattanDistance — Method.

ManhattanDistance(X, Y)

Returns the Manhattan distance matrix of X and Y such that the columns are the samples in Y.

source

ChemometricsTools.ManhattanDistance — Method.

ManhattanDistance(X)

Returns the Manhattan distance matrix of X.

source

ChemometricsTools.Mean — Method.

Mean(rv::RunningMean)

Returns the current mean inside of a RunningMean object.

source

ChemometricsTools.Mean — Method.

Mean(rv::RunningVar)

Returns the current mean inside of a RunningVar object.

source

ChemometricsTools.NMF — Method.

NMF(X; Factors = 1, tolerance = 1e-7, maxiters = 200)

Performs a variation of non-negative matrix factorization on Array X and returns the a 2-Tuple of (Concentration Profile, Spectra)

Note: This is not a coordinate descent based NMF. This is a simple fast version which works well enough for chemical signals Algorithms for non-negative matrix factorization. Daniel D. Lee. H. Sebastian Seung. NIPS'00 Proceedings of the 13th International Conference on Neural Information Processing Systems. 535-54

source

ChemometricsTools.OneHotOdds — Method.

OneHotOdds(Y)

Calculates the odds of a one-hot formatted probability matrix. Returns a tuple.

source

ChemometricsTools.PCA_NIPALS — Method.

PCA_NIPALS(X; Factors = minimum(size(X)) - 1, tolerance = 1e-7, maxiters = 200)

Compute's a PCA from x using the NIPALS algorithm with a user specified number of latent variables(Factors). The tolerance is the minimum change in the F norm before ceasing execution. Returns a PCA object.

source

ChemometricsTools.PSO — Method.

PSO(fn, Bounds, VelRange, Particles; tolerance = 1e-6, maxiters = 1000, InertialDecay = 0.5, PersonalWeight = 0.5, GlobalWeight = 0.5, InternalParams = nothing)

Minimizes function fn with-in the user specified Bounds via a Particle Swarm Optimizer. The particle velocities are limitted to the VelRange. The number of particles are defined by the Particles parameter.

Returns a Tuple of the following form: ( GlobalBestPos, GlobalBestScore, P ) Where P is an array of the particles used in the optimization.

*Note: if the optimization function requires an additional constant parameter, please pass that parameter to InternalParams. This will only work if the optimized parameter(o) and constant parameter(c) for the function of interest has the following format: F(o,c) *

source

ChemometricsTools.PearsonCorrelationCoefficient — Method.

PearsonCorrelationCoefficient( y, yhat )

Calculates The Pearson Correlation Coefficient from vectors Y and YHat

source

ChemometricsTools.PercentRMSE — Method.

PercentRMSE( y, yhat )

Calculates Percent Root Mean Squared Error from vectors Y and YHat

source

ChemometricsTools.PerfectSmoother — Method.

PerfectSmoother(y; lambda = 100)

Applies an assymetric least squares smoothing function to a vector y. The lambda parameter controls the smoothness. See the reference below for more information.

Paul H. C. Eilers. "A Perfect Smoother". Analytical Chemistry, 2003, 75 (14), pp 3631–3636.

source

ChemometricsTools.Q — Method.

Q(X, pca::PCA; Quantile = 0.95, Variance = 1.0)

Computes the Q-statistic and upper control limit cut off of a pca object using a specified Quantile and cumulative variance explained Variance for new or old data X.

A review of PCA-based statistical process monitoring methodsfor time-dependent, high-dimensional data. Bart De Ketelaere https://wis.kuleuven.be/stat/robust/papers/2013/deketelaere-review.pdf

source

ChemometricsTools.RMSE — Method.

RMSE( y, yhat )

Calculates Root Mean Squared Error from vectors Y and YHat

source

ChemometricsTools.RSquare — Method.

RSquare( y, yhat )

Calculates R^2 from Y and YHat

source

ChemometricsTools.Remove! — Method.

Remove!(RM::RunningMean, x)

Removes an observation(x) from a RunningMean object(RM) and reculates the mean in place.

source

ChemometricsTools.Remove — Method.

Remove!(RM::RunningMean, x)

Removes an observation(x) from a RunningMean object(RM) and recuturns the new RunningMean object.

source

ChemometricsTools.SIMPLISMA — Method.

SIMPLISMA(X; Factors = 1)

Performs SIMPLISMA on Array X. Returns a tuple of the following form: (Concentraion Profile, Pure Spectral Estimates, Pure Variables)

Note: This is not the traditional SIMPLISMA algorithm presented by Willem Windig. REAL-TIME WAVELET COMPRESSION AND SELF-MODELING CURVE RESOLUTION FOR ION MOBILITY SPECTROMETRY. PhD. Dissertation. 2003. Guoxiang Chen.

source

ChemometricsTools.SSE — Method.

SSE( y, yhat )

Calculates Sum of Squared Errors from vectors Y and YHat

source

ChemometricsTools.SSReg — Method.

SSReg( y, yhat )

Calculates Sum of Squared Deviations due to Regression from vectors Y and YHat

source

ChemometricsTools.SSRes — Method.

SSRes( y, yhat )

Calculates Sum of Squared Residuals from vectors Y and YHat

source

ChemometricsTools.SSTotal — Method.

SSTotal( y, yhat )

Calculates Total Sum of Squared Deviations from vectors Y and YHat

source

ChemometricsTools.SavitzkyGolay — Method.

SavitzkyGolay(X, Delta, PolyOrder, windowsize)

Performs SavitskyGolay smoothing across every row in an Array X. The window size is the size of the convolution filter, PolyOrder is the order of the polynomial, and Delta is the order of the derivative.

Savitzky, A.; Golay, M.J.E. (1964). "Smoothing and Differentiation of Data by Simplified Least Squares Procedures". Analytical Chemistry. 36 (8): 1627–39. doi:10.1021/ac60214a047.

source

ChemometricsTools.Scale1Norm — Method.

Scale1Norm(X)

Scales the columns of X by the 1-Norm of each row. Returns the scaled array.

source

ChemometricsTools.Scale2Norm — Method.

Scale2Norm(X)

Scales the columns of X by the 2-Norm of each row. Returns the scaled array.

source

ChemometricsTools.ScaleInfNorm — Method.

ScaleInfNorm(X)

Scales the columns of X by the Inf-Norm of each row. Returns the scaled array.

source

ChemometricsTools.SecondDerivative — Method.

FirstDerivative(X)

Uses the finite difference method to compute the second derivative for every row in X. Note: This operation results in the loss of two columns.

source

ChemometricsTools.Shuffle! — Method.

Shuffle!( X, Y )

Shuffles the rows of the X and Y data without replacement in place. In place, means that this function alters the order of the data in memory and this function does not return anything.

source

ChemometricsTools.Shuffle — Method.

Shuffle( X, Y )

Shuffles the rows of the X and Y data without replacement. It returns a 2-Tuple of the shuffled set.

source

ChemometricsTools.SplitByProportion — Function.

SplitByProportion(X::Array, Proportion::Float64 = 0.5)

Splits X Array along the observations dimension into a 2-Tuple based on the Proportion. The form of the output is the following: ( X1, X2 )

source

ChemometricsTools.SplitByProportion — Function.

SplitByProportion(X::Array, Y::Array,Proportion::Float64 = 0.5)

Splits an X and Associated Y Array along the observations dimension into a 2-Tuple of 2-Tuples based on the Proportion. The form of the output is the following: ( (X1, Y1), (X2, Y2) )

source

ChemometricsTools.SquareEuclideanDistance — Method.

SquareEuclideanDistance(X, Y)

Returns the squared euclidean distance matrix of X and Y such that the columns are the samples in Y.

source

ChemometricsTools.SquareEuclideanDistance — Method.

SquareEuclideanDistance(X)

Returns the squared Grahm aka the euclidean distance matrix of X.

source

ChemometricsTools.StandardNormalVariate — Method.

StandardNormalVariate(X)

Scales the columns of X by the mean and standard deviation of each row. Returns the scaled array.

source

ChemometricsTools.Update! — Method.

Update!(RM::RunningMean, x)

Adds new observation(x) to a RunningMean object(RM) in place.

source

ChemometricsTools.Update! — Method.

Update!(RV::RunningVar, x)

Adds new observation(x) to a RunningVar object(RV) and updates it in place.

source

ChemometricsTools.Update — Method.

Update!(RM::RunningMean, x)

Adds new observation(x) to a RunningMean object(RM) and returns the new object.

source

ChemometricsTools.Variance — Method.

Variance(P::ewma)

This function returns the EWMA control variance.

source

ChemometricsTools.Variance — Method.

Variance(rv::RunningVar)

Returns the current variance inside of a RunningVar object.

source

ChemometricsTools.VenetianBlinds — Method.

VenetianBlinds(X,Y)

Splits an X and associated Y Array along the observation dimension into a 2-Tuple of 2-Tuples based on the whether it is even or odd. The form of the output is the following: ( (X1,Y1), (X2, Y2) )

source

ChemometricsTools.VenetianBlinds — Method.

VenetianBlinds(X)

Splits an X Array along the observations dimension into a 2-Tuple of 2-Tuples based on the whether it is even or odd. The form of the output is the following: ( X1, X2 )

source

ChemometricsTools.entropy — Method.

entropy(v)

Calculates the Shannon-Entropy of a probability vector v. Returns a scalar. A common gain function used in tree methods.

source

ChemometricsTools.findpeaks — Method.

findpeaks( vY; m = 3)

Finds the indices of peaks in a vector vY with a window span of 2m. Original R function by Stas_G:(https://stats.stackexchange.com/questions/22974/how-to-find-local-peaks-valleys-in-a-series-of-data) This version is based on a C++ variant by me.

source

ChemometricsTools.gini — Method.

gini(p)

Calculates the GINI coefficient of a probability vector p. Returns a scalar. A common gain function used in tree methods.

source

ChemometricsTools.offsetToZero — Method.

offsetToZero(X)

Ensures that no observation(row) of Array X is less than zero, by ensuring the minimum value of each row is zero.

source

ChemometricsTools.plotchem — Method.

plotchem(QQ::{QQ, BlandAltman}; title )

returns either a QQ Plot or a Bland-Altman plot with the defined title

source

ChemometricsTools.rbinomial — Method.

rbinomial( p, size... )

Makes an N-dimensional array of size(s) size with a probability of being a 1 over a 0 of 1 p.

source

ChemometricsTools.ssd — Method.

ssd(p)

Calculates the sum squared deviations from a decision tree split. Accepts a vector of values, and the mean of that vector. Returns a scalar. A common gain function used in tree methods.

source

ChemometricsTools.stackedweights — Method.

stackedweights(ErrVec; power = 2)

Weights stacked interval errors by the reciprocal power specified. Used for SIPLS, SISPLS, etc.

Ni, W. , Brown, S. D. and Man, R. (2009), Stacked partial least squares regression analysis for spectral calibration and prediction. J. Chemometrics, 23: 505-517. doi:10.1002/cem.1246

source

ChemometricsTools.DirectStandardizationXform — Method.

(DSX::DirectStandardizationXform)(X; Factors = length(DSX.pca.Values))

Applies a the transform from a learned direct standardization object DSX to new data X.

source

ChemometricsTools.ewma — Method.

EWMA(P::ewma)(New; train = true)

Provides an EWMA score for a New scalar value. If train == true the model is updated to include this new value.

source

ChemometricsTools.ChangeCenter — Method.

ChangeCenter(P::ewma, new::Float64)

This is a convenience function to update the center of a P EWMA model, to a new scalar value.

source

ChemometricsTools.ScaleMinMax — Method.

ScaleMinMax(X)

Scales the columns of X by the Min and Max of each row such that no observation is greater than 1 or less than zero. Returns the scaled array.

source

ChemometricsTools.boxcar — Method.

boxcar(X; windowsize = 3, fn = mean)

Applies a boxcar function (fn) to each window of size windowsize to every row in X.

source