Anomaly Detection

Anomaly Detection API Reference

ChemometricsTools has a few anomaly detection methods. Feel free to read the API below. If that's too abstract, check out the shoot-out example : AnomalyDetection

Functions

ChemometricsTools.Hotelling — Method.

Hotelling(X, pca::PCA; Quantile = 0.05, Variance = 1.0)

Computes the hotelling Tsq and upper control limit cut off of a pca object using a specified Quantile and cumulative variance explained Variance for new or old data X.

A review of PCA-based statistical process monitoring methodsfor time-dependent, high-dimensional data. Bart De Ketelaere https://wis.kuleuven.be/stat/robust/papers/2013/deketelaere-review.pdf

ChemometricsTools.Leverage — Method.

Leverage(pca::PCA)

Calculates the leverage of samples in a pca object.

ChemometricsTools.Q — Method.

Q(X, pca::PCA; Quantile = 0.95, Variance = 1.0)

Computes the Q-statistic and upper control limit cut off of a pca object using a specified Quantile and cumulative variance explained Variance for new or old data X.

A review of PCA-based statistical process monitoring methodsfor time-dependent, high-dimensional data. Bart De Ketelaere https://wis.kuleuven.be/stat/robust/papers/2013/deketelaere-review.pdf