Ensemble Models

Ensemble Models API Reference

Functions

ChemometricsTools.RandomForest — Type.

RandomForest(x, y, mode = :classification; gainfn = entropy, trees = 50, maxdepth = 10,  minbranchsize = 5, samples = 0.7, maxvars = nothing)

Returns a classification (mode = :classification) or a regression (mode = :regression) random forest model. The gainfn can be entropy or gini for classification or ssd for regression. If the number of maximumvars is not provided it will default to sqrt(variables) for classification or variables/3 for regression.

The returned object can be used for inference by calling new data on the object as a function.

Breiman, L. Machine Learning (2001) 45: 5. https://doi.org/10.1023/A:1010933404324

ChemometricsTools.RandomForest — Method.

(RF::RandomForest)(X)

Returns bagged prediction vector of random forest model.

ChemometricsTools.MakeIntervals — Function.

MakeIntervals( columns::Int, intervalsize::Union{Array, Tuple} = [20, 50, 100] )

Creates an Dictionary whose key is the interval size and values are an array of intervals from the range: 1 - columns of size intervalsize.

ChemometricsTools.MakeIntervals — Function.

MakeIntervals( columns::Int, intervalsize::Int = 20 )

Returns an 1-Array of intervals from the range: 1 - columns of size intervalsize.

ChemometricsTools.stackedweights — Method.

stackedweights(ErrVec; power = 2)

Weights stacked interval errors by the reciprocal power specified. Used for SIPLS, SISPLS, etc.

Ni, W. , Brown, S. D. and Man, R. (2009), Stacked partial least squares regression analysis for spectral calibration and prediction. J. Chemometrics, 23: 505-517. doi:10.1002/cem.1246