Full API

API

ChemometricsTools.LDA — Method.

LDA(X, Y; Factors = 1)

Compute's a LinearDiscriminantAnalysis transform from x with a user specified number of latent variables(Factors). Returns an LDA object.

source

ChemometricsTools.LDA — Method.

( model::LDA )( Z; Factors = length(model.Values) )

Calling a LDA object on new data brings the new data Z into the LDA basis.

source

ChemometricsTools.PCA — Method.

PCA(X; Factors = minimum(size(X)) - 1)

Compute's a PCA from x using LinearAlgebra's SVD algorithm with a user specified number of latent variables(Factors). Returns a PCA object.

source

ChemometricsTools.PCA — Method.

(T::PCA)(Z::Array; Factors = length(T.Values), inverse = false)

Calling a PCA object on new data brings the new data Z into or out of (inverse = true) the PCA basis.

source

ChemometricsTools.ColdToHot — Method.

ColdToHot(Y, Schema::ClassificationLabel)

Turns a cold encoded Y vector into a one hot encoded array.

source

ChemometricsTools.ExplainedVariance — Method.

ExplainedVariance(lda::LDA)

Calculates the explained variance of each singular value in an LDA object.

source

ChemometricsTools.ExplainedVariance — Method.

ExplainedVariance(PCA::PCA)

Calculates the explained variance of each singular value in a pca object.

source

ChemometricsTools.HighestVote — Method.

HighestVote(yhat)

Returns the column index for each row that has the highest value in one hot encoded yhat. Returns a one cold encoded vector.

source

ChemometricsTools.HighestVoteOneHot — Method.

HighestVoteOneHot(yhat)

Turns the highest column-wise value to a 1 and the others to zeros per row in a one hot encoded yhat. Returns a one cold encoded vector.

source

ChemometricsTools.HotToCold — Method.

HotToCold(Y, Schema::ClassificationLabel)

Turns a one hot encoded Y array into a cold encoded vector.

source

ChemometricsTools.IsColdEncoded — Method.

IsColdEncoded(Y)

Returns a boolean true if the array Y is cold encoded, and false if not.

source

ChemometricsTools.LabelEncoding — Method.

" LabelEncoding(HotOrCold)

Determines if an Array, Y, is one hot encoded, or cold encoded by it's dimensions. Returns a ClassificationLabel object/schema to convert between the formats.

source

ChemometricsTools.MAE — Method.

MAE( y, yhat )

Calculates Mean Average Error from vectors Y and YHat

source

ChemometricsTools.MAPE — Method.

MAPE( y, yhat )

Calculates Mean Average Percent Error from vectors Y and YHat

source

ChemometricsTools.ME — Method.

ME( y, yhat )

Calculates Mean Error from vectors Y and YHat

source

ChemometricsTools.MSE — Method.

MSE( y, yhat )

Calculates Mean Squared Error from vectors Y and YHat

source

ChemometricsTools.MulticlassStats — Method.

MulticlassStats(Y, GT, schema; Microaverage = true)

Calculates many essential classification statistics based on predicted values Y, and ground truth values GT, using the encoding schema. Returns a dictionary of many statistics...

source

ChemometricsTools.MulticlassThreshold — Method.

MulticlassThreshold(yhat; level = 0.5)

Effectively does the same thing as Threshold() but per-row across columns.

Warning this function can allow for no class assignments. HighestVote is preferred

source

ChemometricsTools.PCA_NIPALS — Method.

PCA_NIPALS(X; Factors = minimum(size(X)) - 1, tolerance = 1e-7, maxiters = 200)

Compute's a PCA from x using the NIPALS algorithm with a user specified number of latent variables(Factors). The tolerance is the minimum change in the F norm before ceasing execution. Returns a PCA object.

source

ChemometricsTools.PearsonCorrelationCoefficient — Method.

PearsonCorrelationCoefficient( y, yhat )

Calculates The Pearson Correlation Coefficient from vectors Y and YHat

source

ChemometricsTools.PercentRMSE — Method.

PercentRMSE( y, yhat )

Calculates Percent Root Mean Squared Error from vectors Y and YHat

source

ChemometricsTools.RMSE — Method.

RMSE( y, yhat )

Calculates Root Mean Squared Error from vectors Y and YHat

source

ChemometricsTools.RSquare — Method.

RSquare( y, yhat )

Calculates R^2 from Y and YHat

source

ChemometricsTools.SSE — Method.

SSE( y, yhat )

Calculates Sum of Squared Errors from vectors Y and YHat

source

ChemometricsTools.SSReg — Method.

SSReg( y, yhat )

Calculates Sum of Squared Deviations due to Regression from vectors Y and YHat

source

ChemometricsTools.SSRes — Method.

SSRes( y, yhat )

Calculates Sum of Squared Residuals from vectors Y and YHat

source

ChemometricsTools.SSTotal — Method.

SSTotal( y, yhat )

Calculates Total Sum of Squared Deviations from vectors Y and YHat

source

ChemometricsTools.Threshold — Method.

Threshold(yhat; level = 0.5)

For a binary vector yhat this decides if the label is a 0 or a 1 based on it's value relative to a threshold level.

source

ChemometricsTools.findpeaks — Method.

findpeaks( vY; m = 3)

Finds the indices of peaks in a vector vY with a window span of 2m. Original R function by Stas_G:(https://stats.stackexchange.com/questions/22974/how-to-find-local-peaks-valleys-in-a-series-of-data) This version is based on a C++ variant by me.

source