Regression Models

Regression Models API Reference

Functions

ChemometricsTools.ClassicLeastSquares — Method.

ClassicLeastSquares( X, Y; Bias = false )

Makes a ClassicLeastSquares regression model of the form Y = AX with or without a Bias term. Returns a CLS object.

source

ChemometricsTools.ClassicLeastSquares — Method.

(M::ClassicLeastSquares)(X)

Makes an inference from X using a ClassicLeastSquares object.

source

ChemometricsTools.LSSVM — Method.

LSSVM( X, Y, Penalty; KernelParameter = 0.0, KernelType = "linear" )

Makes a LSSVM model of the form Y = AK with a bias term using a user specified Kernel("linear", or "gaussian") and has an L2 Penalty. Returns a LSSVM Wrapper for a CLS object.

source

ChemometricsTools.LSSVM — Method.

(M::LSSVM)(X)

Makes an inference from X using a LSSVM object.

source

ChemometricsTools.PartialLeastSquares — Method.

PartialLeastSquares( X, Y; Factors = minimum(size(X)) - 2, tolerance = 1e-8, maxiters = 200 )

Returns a PartialLeastSquares regression model object from arrays X and Y.

PARTIAL LEAST-SQUARES REGRESSION: A TUTORIAL PAUL GELADI and BRUCE R.KOWALSKI. Analytica Chimica Acta, 186, (1986) PARTIAL LEAST-SQUARES REGRESSION:

Martens H., NÊs T. Multivariate Calibration. Wiley: New York, 1989.

Re-interpretation of NIPALS results solves PLSR inconsistency problem. Rolf Ergon. Published in Journal of Chemometrics 2009; Vol. 23/1: 72-75

source

ChemometricsTools.PartialLeastSquares — Method.

(M::PartialLeastSquares)

Makes an inference from X using a PartialLeastSquares object.

source

ChemometricsTools.PrincipalComponentRegression — Method.

(M::PrincipalComponentRegression)( X )

Makes an inference from X using a PrincipalComponentRegression object.

source

ChemometricsTools.PrincipalComponentRegression — Method.

PrincipalComponentRegression(PCAObject, Y )

Makes a PrincipalComponentRegression model object from a PCA Object and property value Y.

source

ChemometricsTools.RidgeRegression — Method.

RidgeRegression( X, Y, Penalty; Bias = false )

Makes a RidgeRegression model of the form Y = AX with or without a Bias term and has an L2 Penalty. Returns a CLS object.

source

ChemometricsTools.RidgeRegression — Method.

(M::RidgeRegression)(X)

Makes an inference from X using a RidgeRegression object which wraps a ClassicLeastSquares object.

source

ChemometricsTools.ExtremeLearningMachine — Function.

ExtremeLearningMachine(X, Y, ReservoirSize = 10; ActivationFn = sigmoid)

Returns a ELM regression model object from arrays X and Y, with a user specified ReservoirSize and ActivationFn.

Extreme learning machine: a new learning scheme of feedforward neural networks. Guang-Bin Huang ; Qin-Yu Zhu ; Chee-Kheong Siew. 2004 IEEE International Joint...

source

ChemometricsTools.KernelRidgeRegression — Method.

KernelRidgeRegression( X, Y, Penalty; KernelParameter = 0.0, KernelType = "linear" )

Makes a KernelRidgeRegression model of the form Y = AK using a user specified Kernel("Linear", or "Guassian") and has an L2 Penalty. Returns a KRR Wrapper for a CLS object.

source

ChemometricsTools.MonotoneRegression — Function.

MonotoneRegression(x, w = nothing)

Performs a monotone/isotonic regression on a vector x. This can be weighted with a vector w.

Exceedingly Simple Monotone Regression. Jan de Leeuw. Version 02, March 30, 2017

source

ChemometricsTools.OrdinaryLeastSquares — Method.

OrdinaryLeastSquares( X, Y; Bias = false )

Makes a ClassicLeastSquares regression model of the form Y = AX with or without a Bias term. Returns a CLS object. This is a wrapper function for CLS, because most other fields refer to this as OLS.

source

ChemometricsTools.sigmoid — Method.

sigmoid(x)

Applies the sigmoid function to a scalar value X. Returns a scalar. Can be broad-casted over an Array.

source

ChemometricsTools.ELM — Method.

(M::ELM)(X)

Makes an inference from X using a ELM object.

source

ChemometricsTools.KRR — Method.

(M::KRR)(X)

Makes an inference from X using a KRR object which wraps a ClassicLeastSquares object.

source