Stats.

Stats API Reference

Functions

ChemometricsTools.RunningMean — Method.

RunningMean(x)

Constructs a running mean object with an initial scalar value of x.

source

ChemometricsTools.RunningVar — Method.

RunningVar(x)

Constructs a RunningVar object with an initial scalar value of x. Note: RunningVar objects implicitly calculate the running mean.

source

ChemometricsTools.EmpiricalQuantiles — Method.

EmpiricalQuantiles(X, quantiles)

Finds the column-wise quantiles of 2-Array X and returns them in a 2-Array of size quantiles by variables. *Note: This copies the array... Use a subset if memory is the concern. *

source

ChemometricsTools.Mean — Method.

Mean(rv::RunningMean)

Returns the current mean inside of a RunningMean object.

source

ChemometricsTools.Mean — Method.

Mean(rv::RunningVar)

Returns the current mean inside of a RunningVar object.

source

ChemometricsTools.Remove! — Method.

Remove!(RM::RunningMean, x)

Removes an observation(x) from a RunningMean object(RM) and reculates the mean in place.

source

ChemometricsTools.Remove — Method.

Remove!(RM::RunningMean, x)

Removes an observation(x) from a RunningMean object(RM) and recuturns the new RunningMean object.

source

ChemometricsTools.SampleSkewness — Method.

SampleSkewness(X)

returns a measure of skewness for vector X that is corrected for a sample of the population.

Joanes, D. N., and C. A. Gill. 1998. “Comparing Measures of Sample Skewness and Kurtosis”. The Statistician 47(1): 183–189.

source

ChemometricsTools.Skewness — Method.

Skewness(X)

returns a measure of skewness for a population vector X.

Joanes, D. N., and C. A. Gill. 1998. “Comparing Measures of Sample Skewness and Kurtosis”. The Statistician 47(1): 183–189.

source

ChemometricsTools.Update! — Method.

Update!(RM::RunningMean, x)

Adds new observation(x) to a RunningMean object(RM) in place.

source

ChemometricsTools.Update! — Method.

Update!(RV::RunningVar, x)

Adds new observation(x) to a RunningVar object(RV) and updates it in place.

source

ChemometricsTools.Update — Method.

Update!(RM::RunningMean, x)

Adds new observation(x) to a RunningMean object(RM) and returns the new object.

source

ChemometricsTools.Variance — Method.

Variance(rv::RunningVar)

Returns the current variance inside of a RunningVar object.

source

ChemometricsTools.rbinomial — Method.

rbinomial( p, size... )

Makes an N-dimensional array of size(s) size with a probability of being a 1 over a 0 of 1 p.

source