Time Series

Time Series API Reference

Functions

ChemometricsTools.EchoStateNetwork — Type.

EchoStateNetwork(X, Y, Reservoir = 1000;
                    L2 = 1e-8, alpha = 0.25, SpectralRadius = 1.00, Sparsity = 0.99, Noise = -1.00,
                    bias = true, burnin = 0)

Currently untested.

source

ChemometricsTools.RollingWindow — Method.

RollingWindow(samples::Int,windowsize::Int,skip::Int)

Creates a RollingWindow iterator from a number of samples and a static windowsize where every iteration skip steps are skipped. The iterator can be used in for loops to iteratively return indices of a dynamic rolling window.

source

ChemometricsTools.RollingWindow — Method.

RollingWindow(samples::Int,windowsize::Int)

Creates a RollingWindow iterator from a number of samples and a static windowsize. The iterator can be used in for loops to iteratively return indices of a dynamic rolling window.

source

ChemometricsTools.SimpleAverage — Method.

SimpleAverage( univariate::Array )

source

ChemometricsTools.SimpleAverage — Method.

SimpleAverage( univariate )

source

ChemometricsTools.SimpleAverage — Method.

(sa::SimpleAverage)()

Predicts with a naive forecaster model.

source

ChemometricsTools.EWMA — Method.

EWMA(Initial::Float64, Lambda::Float64) = ewma(Lambda, Initial, Initial, RunningVar(Initial))

Constructs an exponentially weighted moving average object from an vector of scalar property values Initial and the decay parameter Lambda. This computes the running statistcs neccesary for creating the EWMA model using the interval provided and updates the center value to the mean of the provided values.

source

ChemometricsTools.EWMA — Method.

EWMA(Initial::Float64, Lambda::Float64) = ewma(Lambda, Initial, Initial, RunningVar(Initial))

Constructs an exponentially weighted moving average object from an initial scalar property value Initial and the decay parameter Lambda. This defaults the center value to be the initial value.

source

ChemometricsTools.Limits — Method.

Limits(P::ewma; k = 3.0)

This function returns the upper and lower control limits with a k span of variance for an EWMA object P.

source

ChemometricsTools.NaiveForecast — Method.

NaiveForecaster( univariate::Array )

Makes a forecaster model that simply predicts the last value it has seen for all future values.

source

ChemometricsTools.Update — Method.

Update( model::naiveforecaster, newdata::Float64 )

source

ChemometricsTools.Variance — Method.

Variance(P::ewma)

This function returns the EWMA control variance.

source

ChemometricsTools.update! — Method.

update!( model::NaiveForecaster, newdata::Float64 )

Makes a forecast model that simply predicts the last value it has seen for all future values.

source

ChemometricsTools.update — Method.

update( model::NaiveForecaster, newdata::Float64 )

returns an updated forecast model that simply predicts the last value it has seen for all future values.

source

ChemometricsTools.NaiveForecaster — Method.

(nf::NaiveForecaster)()

Predicts with a naive forecaster model.

source

ChemometricsTools.ewma — Method.

EWMA(P::ewma)(New; train = true)

Provides an EWMA score for a New scalar value. If train == true the model is updated to include this new value.

source

ChemometricsTools.ChangeCenter — Method.

ChangeCenter(P::ewma, new::Float64)

This is a convenience function to update the center of a P EWMA model, to a new scalar value.

source