Transformations/Pipelines

Transformations/Pipelines API Reference

Functions

ChemometricsTools.Center — Method.

(T::Center)(Z; inverse = false)

Centers data in array Z column-wise according to learned mean centers in Center object T.

source

ChemometricsTools.CenterScale — Method.

(T::CenterScale)(Z; inverse = false)

Centers and Scales data in array Z column-wise according to learned measures of central tendancy in Scale object T.

source

ChemometricsTools.QuantileTrim — Type.

(T::QuantileTrim)(X, inverse = false)

Trims data in array X columns wise according to learned quantiles in QuantileTrim object T This function does NOT have an inverse.

source

ChemometricsTools.QuantileTrim — Method.

QuantileTrim(Z; quantiles::Tuple{Float64,Float64} = (0.05, 0.95) )

Trims values above or below the specified columnwise quantiles to the quantile values themselves.

source

ChemometricsTools.RangeNorm — Method.

(T::RangeNorm)(Z; inverse = false)

Scales and shifts data in array Z column-wise according to learned min-maxes in RangeNorm object T.

source

ChemometricsTools.Scale — Method.

(T::Scale)(Z; inverse = false)

Scales data in array Z column-wise according to learned standard deviations in Scale object T.

source

ChemometricsTools.Logit — Method.

Logit(Z; inverse = false)

Logit transforms (ln( X / (1 - X) ))) every element in Z. The inverse may also be applied. Warning: This can return Infs and NaNs if elements of Z are not suited to the transform

source

ChemometricsTools.Pipeline — Method.

Pipeline( X, FnStack... )

Construct a pipeline object from vector/tuple of Transforms. The Transforms vector are effectively a vector of functions which transform data.

source

ChemometricsTools.Pipeline — Method.

Pipeline(Transforms)

Constructs a transformation pipeline from vector/tuple of Transforms. The Transforms vector are effectively a vector of functions which transform data.

source

ChemometricsTools.PipelineInPlace — Method.

PipelineInPlace( X, FnStack...)

Construct a pipeline object from vector/tuple of Transforms. The Transforms vector are effectively a vector of functions which transform data. This function makes "inplace" changes to the Array X as though it has been sent through the pipeline. This is more efficient if memory is a concern, but can irreversibly transform data in memory depending on the transforms in the pipeline.

source

ChemometricsTools.pipeline — Method.

(P::pipeline)(X; inverse = false)

Applies the stored transformations in a pipeline object P to data in X. The inverse flag can allow for the transformations to be reversed provided they are invertible functions.

source