Anomaly Detection

Anomaly Detection API Reference

ChemometricsTools has a few anomaly detection methods. Feel free to read the API below. If that's too abstract, check out the shoot-out example : AnomalyDetection

Functions

ChemometricsTools.OneClassJKNN — Method.

OneClassJKNN( Normal, New; J::Int = 1, K::Int = 1, DistanceType = "euclidean" )

Creates a one class JK-NN classifier from Normal data and evaluates it on New data. This compares the inter sample distance (DistanceType) between a New and Normal J nearest neighbors to the K nearest neighbors of those J nearest neighbors in the Normal set. No cut off is provided, that should be done by the end-user. A typical cut off value is 1.0 .

source