Distances/Kernels

Distances API Reference

Functions

ChemometricsTools.Kernel — Method.

(K::Kernel)(X)

This is a convenience function to allow for one-line construction of kernels from a Kernel object K and new data X.

source

ChemometricsTools.Kernel — Method.

Kernel(X::Array)

Default constructor for Kernel object. Returns the linear kernel of X.

source

ChemometricsTools.AdjacencyMatrix — Method.

NearestNeighbors(DistanceMatrix)

Returns the nearest neighbor adjacency matrix from a given DistanceMatrix.

source

ChemometricsTools.CauchyKernel — Method.

CauchyKernel(X, Y, sigma)

Creates a Cauchy kernel from Arrays X and Y using hyperparameters sigma.

source

ChemometricsTools.CauchyKernel — Method.

CauchyKernel(X, sigma)

Creates a Cauchy kernel from Array X using hyperparameters sigma.

source

ChemometricsTools.CenterKernelMatrix — Function.

CenterKernelMatrix(X)

Returns a centered kernel matrix.

source

ChemometricsTools.EuclideanDistance — Method.

EuclideanDistance(X, Y)

Returns the euclidean distance matrix of X and Y such that the columns are the samples in Y.

source

ChemometricsTools.EuclideanDistance — Method.

EuclideanDistance(X)

Returns the Grahm aka the euclidean distance matrix of X.

source

ChemometricsTools.GaussianKernel — Method.

GaussianKernel(X, Y, sigma)

Creates a Gaussian/RBF kernel from Arrays X and Y with hyperparameter sigma.

source

ChemometricsTools.GaussianKernel — Method.

GaussianKernel(X, sigma)

Creates a Gaussian/RBF kernel from Array X using hyperparameter sigma.

source

ChemometricsTools.InClassAdjacencyMatrix — Function.

InClassAdjacencyMatrix(DistanceMatrix, YHOT, K = 1)

Computes the in class Adjacency matrix with K nearest neighbors.

source

ChemometricsTools.LevenshteinDistance — Method.

LevenshteinDistance(s::AbstractString, t::AbstractString)

Calculates the LevenshteinDistance aka the edit distance between 2 strings.

Borrowed from: https://rosettacode.org/wiki/Levenshtein_distance#Julia

source

ChemometricsTools.LinearKernel — Method.

LinearKernel(X, Y, c)

Creates a Linear kernel from Arrays X and Y with hyperparameter C.

source

ChemometricsTools.LinearKernel — Method.

LinearKernel(X, c)

Creates a Linear kernel from Array X and hyperparameter C.

source

ChemometricsTools.ManhattanDistance — Method.

ManhattanDistance(X, Y)

Returns the Manhattan distance matrix of X and Y such that the columns are the samples in Y.

source

ChemometricsTools.ManhattanDistance — Method.

ManhattanDistance(X)

Returns the Manhattan distance matrix of X.

source

ChemometricsTools.MinkowskiDistance — Method.

MinkowskiDistance(X, Y, p)

Returns the Minkowski distance matrix of X and Y using order p such that the columns are the samples in Y.

source

ChemometricsTools.MinkowskiDistance — Method.

MinkowskiDistance(X, p)

Returns the Manhattan distance matrix of X using order p.

source

ChemometricsTools.NearestNeighbors — Method.

NearestNeighbors(DistanceMatrix, N)

Returns a matrix of dimensions DistanceMatrix rows, by N columns. Basically this goes through each row and finds the ones corresponding column which has the smallest distance.

source

ChemometricsTools.OutOfClassAdjacencyMatrix — Function.

OutOfClassAdjacencyMatrix(DistanceMatrix, YHOT, K = 1)

Computes the out of class Adjacency matrix with K nearest neighbors.

source

ChemometricsTools.SquareEuclideanDistance — Method.

SquareEuclideanDistance(X, Y)

Returns the squared euclidean distance matrix of X and Y such that the columns are the samples in Y.

source

ChemometricsTools.SquareEuclideanDistance — Method.

SquareEuclideanDistance(X)

Returns the squared Grahm aka the euclidean distance matrix of X.

source