Plotting

Plotting Tools API Reference

Functions

ChemometricsTools.BlandAltman — Method.

BlandAltman(Y1, Y2; Confidence = 1.96)

Returns a Plottable object of a Bland-Altman plot between vectors Y1 and Y2 with a confidence limit of Confidence.

source

ChemometricsTools.DAPlot — Type.

DiscriminantAnalysisPlot(DA, GD, YHot, LblEncoding, UnlabeledData,
                            Axis = [1,2], Confidence = 0.90)

An object to hold all the information needed to plot a discriminant analysis. DA = LDA/PCA object GD = GuassianDiscriminant object YHot = One hot encoded y vectors LblEncoding = a ClassificationLabel object UnlabeledData = data that does not have a known label (see where it falls in the plot). Axis = principal or discriminant axis' to display Confidence = Significance value from 0-1 for the confidence ellipses.

source

ChemometricsTools.IntervalOverlay — Type.

IntervalOverlay(Spectra, Intervals, Err)

Creates a Plottable object to display the relative error(Err) of each interval ontop of a Spectra.

source

ChemometricsTools.QQ — Method.

QQ( Y1, Y2; Quantiles = collect( 1 : 99 ) ./ 100 )

Returns a plotable object of a Quantile-Quantile plot between vectors Y1 and Y2 at the desired Quantiles.

source

RecipesBase.apply_recipe — Method.

blandaltman(BA::BlandAltman)

Plots a Bland-Altman plot.

source

RecipesBase.apply_recipe — Method.

residualsplotrecipe(rmodel::ChemometricsTools.RegressionModels, X, Y)

Plots the residuals of a RegressionModel object given an input X and ground-truth Y.

source

RecipesBase.apply_recipe — Method.

discrimanalyplot(dap::DAPlot)

Plots two DA axis, with ellipsoids according to a DiscriminantAnalysisPlot object.

source

RecipesBase.apply_recipe — Method.

intervaloverlay(IO::IntervalOverlay)

Plots a barchart overlay over a spectra according to an IntervalOverlay object.

source

RecipesBase.apply_recipe — Method.

quantilequantile(qq::QQ)

Plots a Quantile-Quantile plot.

source

RecipesBase.apply_recipe — Method.

standard_addition_recipe(rmodel::StandardAddition, X, Y;
                                        decimals = 3, text_location = :topleft)

Plots a trend line of a StandardAddition object given an input X and ground-truth Y.

source

RecipesBase.apply_recipe — Method.

PCA/LDA(::Union{PCA, LDA}; Axis = [1,2])

Plots scores of PCA/LDA object using the defined Axis'.

source

RecipesBase.apply_recipe — Method.

univariatecalibrationrecipe(rmodel::UnivariateCalibration, X, Y;
                                        decimals = 3, text_location = :topleft)

Plots the calibration trend line of a UnivariateCalibration object given an input X and ground-truth Y.

source