Univariate

Univariate Methods API Reference

Functions

ChemometricsTools.StandardAddition — Method.

StandardAddition( Signal, Spike )

Performs a univariate standard addition calibration. Returns a StandardAddition

object.

source

ChemometricsTools.UnivariateCalibration — Method.

UnivariateCalibration( Y, X )

Performs a univariate least squares regression with a bias/offset term(`y = mx + b`).

Returns a Univariate Calibration object containing uncertainty in parameters, and other statistics.

source

ChemometricsTools.Confidence_Offset — Method.

Confidence_Offset( UC::Univariate; Significance = 0.05 )

Returns a tuple of (-, +) of the estimated offset/bias's confidence interval from a UC Univariate type object.

source

ChemometricsTools.Confidence_Slope — Method.

Confidence_Slope( UC::Univariate; Significance = 0.05 )

Returns a tuple of (-, +) of the estimated slope's confidence interval from a UC Univariate type object.

source