Data Utilities

Data Utilities API Reference

Functions

ChemometricsTools.ChemometricsToolsDataset — Method.

ChemometricsToolsDataset(file::Int)

Loads in a dataset included in the ChemometricsTools.jl package.

source

ChemometricsTools.ChemometricsToolsDataset — Method.

ChemometricsToolsDataset(file::Int)

Loads in a dataset included in the ChemometricsTools.jl package.

source

ChemometricsTools.ChemometricsToolsDatasets — Method.

ChemometricsToolsDatasets()

Displays a list of the available datasets in ChemometricsTools.jl .

source

ChemometricsTools.FindCommonVariables — Method.

FindCommonVariables( DFArray::Array{DataFrame,1} )

Finds exact matches across variable names over an array of DataFrames. Variables that are not mutual to each dataframe are discarded.

For intelligent/soft matching please see ExploreCommonVariables().

source