Kernel Density Generator

Kernel Density Generator API Reference

Functions

ChemometricsTools.GaussianBand — Type.

GaussianBand(sigma,amplitude,center)

Constructs a Gaussian kernel generator.

source

ChemometricsTools.GaussianBand — Method.

(B::GaussianBand)(X::Float64)

Returns the scalar probability associated with a GaussianBand object (kernel) at a location in space(X).

source

ChemometricsTools.LorentzianBand — Type.

LorentzianBand(gamma,amplitude,center)

Constructs a Lorentzian kernel generator.

source

ChemometricsTools.LorentzianBand — Method.

(B::LorentzianBand)(X::Float64)

Returns the probability associated with a LorentzianBand object (kernel) at a location in space(X).

source

ChemometricsTools.Universe — Method.

(U::Universe)(Band...)

A Universe objects internal "spectra" can be updated to include the additive contribution of many Band-like objects.

source

ChemometricsTools.Universe — Method.

Universe(mini, maxi; width = nothing, bins = nothing)

Creates a 1-D discretized segment that starts at mini and ends at maxi. The width of the bins for the discretization can either be provided or inferred from the number of bins. Returns a Universe object.

source

ChemometricsTools.Universe — Method.

(U::Universe)(Band::Union{ GaussianBand, LorentzianBand})

A Universe objects internal "spectra" can be updated to include the additive contribution of any Band-like object.

source

ChemometricsTools.SpectralArray — Method.

SpectralArray(Universes::Array{Universe,1})

Takes an array of Universe types and returns a 2-Array of the spectra.

source