Regression Metrics

Regression Metrics API Reference

Functions

ChemometricsTools.FNorm — Method.

FNorm( X )

Calculates the Froebinius norm of matrix X.

source

ChemometricsTools.MAE — Method.

MAE( y, yhat )

Calculates Mean Average Error from vectors Y and YHat

source

ChemometricsTools.MAPE — Method.

MAPE( y, yhat )

Calculates Mean Average Percent Error from vectors Y and YHat

source

ChemometricsTools.ME — Method.

ME( y, yhat )

Calculates Mean Error from vectors Y and YHat.

source

ChemometricsTools.MSE — Method.

MSE( y, yhat )

Calculates Mean Squared Error from vectors Y and YHat

source

ChemometricsTools.PearsonCorrelationCoefficient — Method.

PearsonCorrelationCoefficient( y, yhat )

Calculates The Pearson Correlation Coefficient from vectors Y and YHat

source

ChemometricsTools.PercentRMSE — Method.

PercentRMSE( y, yhat; aspercent = true )

Calculates Percent Root Mean Squared Error from vectors Y and YHat

source

ChemometricsTools.RMSE — Method.

RMSE( y, yhat )

Calculates Root Mean Squared Error from vectors Y and YHat

source

ChemometricsTools.RSquare — Method.

RSquare( y, yhat )

Calculates R^2 from Y and YHat

source

ChemometricsTools.SSE — Method.

SSE( y, yhat )

Calculates Sum of Squared Errors from vectors Y and YHat

source

ChemometricsTools.SSReg — Method.

SSReg( y, yhat )

Calculates Sum of Squared Deviations due to Regression from vectors Y and YHat

source

ChemometricsTools.SSRes — Method.

SSRes( y, yhat )

Calculates Sum of Squared Residuals from vectors Y and YHat

source

ChemometricsTools.SSTotal — Method.

SSTotal( y, yhat )

Calculates Total Sum of Squared Deviations from vectors Y and YHat

source