Documentation

Documentation

ChemometricsTools.KFoldsValidation — Method.

KFoldsValidation(K::Int, x, y)

Returns a KFoldsValidation iterator with K folds. Because it's an iterator it can be used in for loops, see the tutorials for pragmatic examples. The iterator returns a 2-Tuple of 2-Tuples which have the following form: ((TrainX,TrainY),(ValidateX,ValidateY).

source

ChemometricsTools.LDA — Method.

LDA(X, Y; Factors = 1)

Compute's a LinearDiscriminantAnalysis transform from x with a user specified number of latent variables(Factors). Returns an LDA object.

source

ChemometricsTools.LDA — Method.

( model::LDA )( Z; Factors = length(model.Values) )

Calling a LDA object on new data brings the new data Z into the LDA basis.

source

ChemometricsTools.PCA — Method.

PCA(X; Factors = minimum(size(X)) - 1)

Compute's a PCA from x using LinearAlgebra's SVD algorithm with a user specified number of latent variables(Factors). Returns a PCA object.

source

ChemometricsTools.PCA — Method.

(T::PCA)(Z::Array; Factors = length(T.Values), inverse = false)

Calling a PCA object on new data brings the new data Z into or out of (inverse = true) the PCA basis.

source

ChemometricsTools.RollingWindow — Method.

RollingWindow(samples::Int,windowsize::Int,skip::Int)

Creates a RollingWindow iterator from a number of samples and a static windowsize where every iteration skip steps are skipped. The iterator can be used in for loops to iteratively return indices of a dynamic rolling window.

source

ChemometricsTools.RollingWindow — Method.

RollingWindow(samples::Int,windowsize::Int)

Creates a RollingWindow iterator from a number of samples and a static windowsize. The iterator can be used in for loops to iteratively return indices of a dynamic rolling window.

source

ChemometricsTools.ClassificationTree — Method.

ClassificationTree(x, y; gainfn = entropy, maxdepth = 4, minbranchsize = 3)

Builds a CART object using either gini or entropy as a partioning method. Y must be a one hot encoded 2-Array. Predictions can be formed by calling the following function from the CART object: (M::CART)(x).

*Note: this is a purely nonrecursive decision tree. The julia compiler doesn't like storing structs of nested things. I wrote it the recursive way in the past and it was quite slow, I think this is true also of interpretted languages like R/Python...So here it is, nonrecursive tree's!

source

ChemometricsTools.EWMA — Method.

EWMA(Initial::Float64, Lambda::Float64) = ewma(Lambda, Initial, Initial, RunningVar(Initial))

Constructs an exponentially weighted moving average object from an vector of scalar property values Initial and the decay parameter Lambda. This computes the running statistcs neccesary for creating the EWMA model using the interval provided and updates the center value to the mean of the provided values.

source

ChemometricsTools.EWMA — Method.

EWMA(Initial::Float64, Lambda::Float64) = ewma(Lambda, Initial, Initial, RunningVar(Initial))

Constructs an exponentially weighted moving average object from an initial scalar property value Initial and the decay parameter Lambda. This defaults the center value to be the initial value.

source

ChemometricsTools.ExplainedVariance — Method.

ExplainedVariance(lda::LDA)

Calculates the explained variance of each singular value in an LDA object.

source

ChemometricsTools.ExplainedVariance — Method.

ExplainedVariance(PCA::PCA)

Calculates the explained variance of each singular value in a pca object.

source

ChemometricsTools.Hotelling — Method.

Hotelling(X, pca::PCA; Quantile = 0.05, Variance = 1.0)

Computes the hotelling Tsq and upper control limit cut off of a pca object using a specified Quantile and cumulative variance explained Variance for new or old data X.

A review of PCA-based statistical process monitoring methodsfor time-dependent, high-dimensional data. Bart De Ketelaere https://wis.kuleuven.be/stat/robust/papers/2013/deketelaere-review.pdf

source

ChemometricsTools.Leverage — Method.

Leverage(pca::PCA)

Calculates the leverage of samples in a pca object.

source

ChemometricsTools.Limits — Method.

Limits(P::ewma; k = 3.0)

This function returns the upper and lower control limits with a k span of variance for an EWMA object P.

source

ChemometricsTools.OneHotOdds — Method.

OneHotOdds(Y)

Calculates the odds of a one-hot formatted probability matrix. Returns a tuple.

source

ChemometricsTools.PCA_NIPALS — Method.

PCA_NIPALS(X; Factors = minimum(size(X)) - 1, tolerance = 1e-7, maxiters = 200)

Compute's a PCA from x using the NIPALS algorithm with a user specified number of latent variables(Factors). The tolerance is the minimum change in the F norm before ceasing execution. Returns a PCA object.

source

ChemometricsTools.Q — Method.

Q(X, pca::PCA; Quantile = 0.95, Variance = 1.0)

Computes the Q-statistic and upper control limit cut off of a pca object using a specified Quantile and cumulative variance explained Variance for new or old data X.

A review of PCA-based statistical process monitoring methodsfor time-dependent, high-dimensional data. Bart De Ketelaere https://wis.kuleuven.be/stat/robust/papers/2013/deketelaere-review.pdf

source

ChemometricsTools.Shuffle! — Method.

Shuffle!( X, Y )

Shuffles the rows of the X and Y data without replacement in place. In place, means that this function alters the order of the data in memory and this function does not return anything.

source

ChemometricsTools.Shuffle — Method.

Shuffle( X, Y )

Shuffles the rows of the X and Y data without replacement. It returns a 2-Tuple of the shuffled set.

source

ChemometricsTools.Variance — Method.

Variance(P::ewma)

This function returns the EWMA control variance.

source

ChemometricsTools.entropy — Method.

entropy(v)

Calculates the Shannon-Entropy of a probability vector v. Returns a scalar. A common gain function used in tree methods.

source

ChemometricsTools.findpeaks — Method.

findpeaks( vY; m = 3)

Finds the indices of peaks in a vector vY with a window span of 2m. Original R function by Stas_G:(https://stats.stackexchange.com/questions/22974/how-to-find-local-peaks-valleys-in-a-series-of-data) This version is based on a C++ variant by me.

source

ChemometricsTools.gini — Method.

gini(p)

Calculates the GINI coefficient of a probability vector p. Returns a scalar. A common gain function used in tree methods.

source

ChemometricsTools.ssd — Method.

ssd(p)

Calculates the sum squared deviations from a decision tree split. Accepts a vector of values, and the mean of that vector. Returns a scalar. A common gain function used in tree methods.

source

ChemometricsTools.ewma — Method.

EWMA(P::ewma)(New; train = true)

Provides an EWMA score for a New scalar value. If train == true the model is updated to include this new value.

source

ChemometricsTools.ChangeCenter — Method.

ChangeCenter(P::ewma, new::Float64)

This is a convenience function to update the center of a P EWMA model, to a new scalar value.

source