Documentation

Documentation

ChemometricsTools.CORAL — Method.

CORAL(X1, X2; lambda = 1.0)

Performs CORAL to facilitate covariance based transfer from X1 to X2 with regularization parameter lambda. Returns a CORAL object.

Correlation Alignment for Unsupervised Domain Adaptation. Baochen Sun, Jiashi Feng, Kate Saenko. https://arxiv.org/abs/1612.01939

source

ChemometricsTools.CORAL — Method.

(C::CORAL)(Z)

Applies a the transform from a learned CORAL object to new data Z.

source

ChemometricsTools.KFoldsValidation — Method.

KFoldsValidation(K::Int, x, y)

Returns a KFoldsValidation iterator with K folds. Because it's an iterator it can be used in for loops, see the tutorials for pragmatic examples. The iterator returns a 2-Tuple of 2-Tuples which have the following form: ((TrainX,TrainY),(ValidateX,ValidateY).

source

ChemometricsTools.LDA — Method.

LDA(X, Y; Factors = 1)

Compute's a LinearDiscriminantAnalysis transform from x with a user specified number of latent variables(Factors). Returns an LDA object.

source

ChemometricsTools.LDA — Method.

( model::LDA )( Z; Factors = length(model.Values) )

Calling a LDA object on new data brings the new data Z into the LDA basis.

source

ChemometricsTools.MultiplicativeScatterCorrection — Method.

(T::MultiplicativeScatterCorrection)(Z)

Applies MultiplicativeScatterCorrection from a stored object T to Array Z.

source

ChemometricsTools.OrthogonalSignalCorrection — Method.

OrthogonalSignalCorrection(X, Y; Factors = 1)

Performs Thomas Fearn's Orthogonal Signal Correction to an endogenous X and exogenous Y. The number of Factors are the number of orthogonal components to be removed from X. This function returns an OSC object.

Tom Fearn. On orthogonal signal correction. Chemometrics and Intelligent Laboratory Systems. Volume 50, Issue 1, 2000, Pages 47-52.

source

ChemometricsTools.OrthogonalSignalCorrection — Method.

(OSC::OrthogonalSignalCorrection)(Z; Factors = 2)

Applies a the transform from a learned orthogonal signal correction object OSC to new data Z.

source

ChemometricsTools.PCA — Method.

PCA(X; Factors = minimum(size(X)) - 1)

Compute's a PCA from x using LinearAlgebra's SVD algorithm with a user specified number of latent variables(Factors). Returns a PCA object.

source

ChemometricsTools.PCA — Method.

(T::PCA)(Z::Array; Factors = length(T.Values), inverse = false)

Calling a PCA object on new data brings the new data Z into or out of (inverse = true) the PCA basis.

source

ChemometricsTools.RollingWindow — Method.

RollingWindow(samples::Int,windowsize::Int,skip::Int)

Creates a RollingWindow iterator from a number of samples and a static windowsize where every iteration skip steps are skipped. The iterator can be used in for loops to iteratively return indices of a dynamic rolling window.

source

ChemometricsTools.RollingWindow — Method.

RollingWindow(samples::Int,windowsize::Int)

Creates a RollingWindow iterator from a number of samples and a static windowsize. The iterator can be used in for loops to iteratively return indices of a dynamic rolling window.

source

ChemometricsTools.RunningMean — Method.

RunningMean(x)

Constructs a running mean object with an initial scalar value of x.

source

ChemometricsTools.RunningVar — Method.

RunningVar(x)

Constructs a RunningVar object with an initial scalar value of x. Note: RunningVar objects implicitly calculate the running mean.

source

ChemometricsTools.TransferByOrthogonalProjection — Method.

TransferByOrthogonalProjection(X1, X2; Factors = 1)

Performs Thomas Fearns Transfer By Orthogonal Projection to facilitate transfer from X1 to X2. Returns a TransferByOrthogonalProjection object.

Anne Andrew, Tom Fearn. Transfer by orthogonal projection: making near-infrared calibrations robust to between-instrument variation. Chemometrics and Intelligent Laboratory Systems. Volume 72, Issue 1, 2004, Pages 51-56,

source

ChemometricsTools.TransferByOrthogonalProjection — Method.

(TbOP::TransferByOrthogonalProjection)(X1; Factors = TbOP.Factors)

Applies a the transform from a learned transfer by orthogonal projection object TbOP to new data X1.

source

ChemometricsTools.ALSSmoother — Method.

ALSSmoother(y; lambda = 100, p = 0.001, maxiters = 10)

Applies an assymetric least squares smoothing function to a vector y. The lambda, p, and maxiters parameters control the smoothness. See the reference below for more information.

Paul H. C. Eilers, Hans F.M. Boelens. Baseline Correction with Asymmetric Least Squares Smoothing. 2005

source

ChemometricsTools.ClassificationTree — Method.

ClassificationTree(x, y; gainfn = entropy, maxdepth = 4, minbranchsize = 3)

Builds a CART object using either gini or entropy as a partioning method. Y must be a one hot encoded 2-Array. Predictions can be formed by calling the following function from the CART object: (M::CART)(x).

*Note: this is a purely nonrecursive decision tree. The julia compiler doesn't like storing structs of nested things. I wrote it the recursive way in the past and it was quite slow, I think this is true also of interpretted languages like R/Python...So here it is, nonrecursive tree's!

source

ChemometricsTools.DirectStandardization — Method.

DirectStandardization(InstrumentX1, InstrumentX2; Factors = minimum(collect(size(InstrumentX1))) - 1)

Makes a DirectStandardization object to facilitate the transfer from Instrument #2 to Instrument #1 . The returned object can be used to transfer unseen data to the approximated space of instrument 1. The number of Factors used are those from the internal orthogonal basis.

Yongdong Wang and Bruce R. Kowalski, "Calibration Transfer and Measurement Stability of Near-Infrared Spectrometers," Appl. Spectrosc. 46, 764-771 (1992)

source

ChemometricsTools.EWMA — Method.

EWMA(Initial::Float64, Lambda::Float64) = ewma(Lambda, Initial, Initial, RunningVar(Initial))

Constructs an exponentially weighted moving average object from an vector of scalar property values Initial and the decay parameter Lambda. This computes the running statistcs neccesary for creating the EWMA model using the interval provided and updates the center value to the mean of the provided values.

source

ChemometricsTools.EWMA — Method.

EWMA(Initial::Float64, Lambda::Float64) = ewma(Lambda, Initial, Initial, RunningVar(Initial))

Constructs an exponentially weighted moving average object from an initial scalar property value Initial and the decay parameter Lambda. This defaults the center value to be the initial value.

source

ChemometricsTools.EmpiricalQuantiles — Method.

EmpiricalQuantiles(X, quantiles)

Finds the column-wise quantiles of 2-Array X and returns them in a 2-Array of size quantiles by variables. *Note: This copies the array... Use a subset if memory is the concern. *

source

ChemometricsTools.ExplainedVariance — Method.

ExplainedVariance(lda::LDA)

Calculates the explained variance of each singular value in an LDA object.

source

ChemometricsTools.ExplainedVariance — Method.

ExplainedVariance(PCA::PCA)

Calculates the explained variance of each singular value in a pca object.

source

ChemometricsTools.FirstDerivative — Method.

FirstDerivative(X)

Uses the finite difference method to compute the first derivative for every row in X. Note: This operation results in the loss of a column dimension.

source

ChemometricsTools.FractionalDerivative — Function.

FractionalDerivative(Y, X = 1 : length(Y); Order = 0.5)

Calculates the Grunwald-Leitnikov fractional order derivative on every row of Array Y. Array X is a vector that has the spacing between column-wise entries in Y. X can be a scalar if that is constant (common in spectroscopy). Order is the fractional order of the derivative. Note: This operation results in the loss of a column dimension.

The Fractional Calculus, by Oldham, K.; and Spanier, J. Hardcover: 234 pages. Publisher: Academic Press, 1974. ISBN 0-12-525550-0

source

ChemometricsTools.Hotelling — Method.

Hotelling(X, pca::PCA; Quantile = 0.05, Variance = 1.0)

Computes the hotelling Tsq and upper control limit cut off of a pca object using a specified Quantile and cumulative variance explained Variance for new or old data X.

A review of PCA-based statistical process monitoring methodsfor time-dependent, high-dimensional data. Bart De Ketelaere https://wis.kuleuven.be/stat/robust/papers/2013/deketelaere-review.pdf

source

ChemometricsTools.Leverage — Method.

Leverage(pca::PCA)

Calculates the leverage of samples in a pca object.

source

ChemometricsTools.Limits — Method.

Limits(P::ewma; k = 3.0)

This function returns the upper and lower control limits with a k span of variance for an EWMA object P.

source

ChemometricsTools.Mean — Method.

Mean(rv::RunningMean)

Returns the current mean inside of a RunningMean object.

source

ChemometricsTools.Mean — Method.

Mean(rv::RunningVar)

Returns the current mean inside of a RunningVar object.

source

ChemometricsTools.OneHotOdds — Method.

OneHotOdds(Y)

Calculates the odds of a one-hot formatted probability matrix. Returns a tuple.

source

ChemometricsTools.PCA_NIPALS — Method.

PCA_NIPALS(X; Factors = minimum(size(X)) - 1, tolerance = 1e-7, maxiters = 200)

Compute's a PCA from x using the NIPALS algorithm with a user specified number of latent variables(Factors). The tolerance is the minimum change in the F norm before ceasing execution. Returns a PCA object.

source

ChemometricsTools.PerfectSmoother — Method.

PerfectSmoother(y; lambda = 100)

Applies an assymetric least squares smoothing function to a vector y. The lambda parameter controls the smoothness. See the reference below for more information.

Paul H. C. Eilers. "A Perfect Smoother". Analytical Chemistry, 2003, 75 (14), pp 3631–3636.

source

ChemometricsTools.Q — Method.

Q(X, pca::PCA; Quantile = 0.95, Variance = 1.0)

Computes the Q-statistic and upper control limit cut off of a pca object using a specified Quantile and cumulative variance explained Variance for new or old data X.

A review of PCA-based statistical process monitoring methodsfor time-dependent, high-dimensional data. Bart De Ketelaere https://wis.kuleuven.be/stat/robust/papers/2013/deketelaere-review.pdf

source

ChemometricsTools.Remove! — Method.

Remove!(RM::RunningMean, x)

Removes an observation(x) from a RunningMean object(RM) and reculates the mean in place.

source

ChemometricsTools.Remove — Method.

Remove!(RM::RunningMean, x)

Removes an observation(x) from a RunningMean object(RM) and recuturns the new RunningMean object.

source

ChemometricsTools.SavitzkyGolay — Method.

SavitzkyGolay(X, Delta, PolyOrder, windowsize)

Performs SavitskyGolay smoothing across every row in an Array X. The window size is the size of the convolution filter, PolyOrder is the order of the polynomial, and Delta is the order of the derivative.

Savitzky, A.; Golay, M.J.E. (1964). "Smoothing and Differentiation of Data by Simplified Least Squares Procedures". Analytical Chemistry. 36 (8): 1627–39. doi:10.1021/ac60214a047.

source

ChemometricsTools.Scale1Norm — Method.

Scale1Norm(X)

Scales the columns of X by the 1-Norm of each row. Returns the scaled array.

source

ChemometricsTools.Scale2Norm — Method.

Scale2Norm(X)

Scales the columns of X by the 2-Norm of each row. Returns the scaled array.

source

ChemometricsTools.ScaleInfNorm — Method.

ScaleInfNorm(X)

Scales the columns of X by the Inf-Norm of each row. Returns the scaled array.

source

ChemometricsTools.SecondDerivative — Method.

FirstDerivative(X)

Uses the finite difference method to compute the second derivative for every row in X. Note: This operation results in the loss of two columns.

source

ChemometricsTools.Shuffle! — Method.

Shuffle!( X, Y )

Shuffles the rows of the X and Y data without replacement in place. In place, means that this function alters the order of the data in memory and this function does not return anything.

source

ChemometricsTools.Shuffle — Method.

Shuffle( X, Y )

Shuffles the rows of the X and Y data without replacement. It returns a 2-Tuple of the shuffled set.

source

ChemometricsTools.StandardNormalVariate — Method.

StandardNormalVariate(X)

Scales the columns of X by the mean and standard deviation of each row. Returns the scaled array.

source

ChemometricsTools.Update! — Method.

Update!(RM::RunningMean, x)

Adds new observation(x) to a RunningMean object(RM) in place.

source

ChemometricsTools.Update! — Method.

Update!(RV::RunningVar, x)

Adds new observation(x) to a RunningVar object(RV) and updates it in place.

source

ChemometricsTools.Update — Method.

Update!(RM::RunningMean, x)

Adds new observation(x) to a RunningMean object(RM) and returns the new object.

source

ChemometricsTools.Variance — Method.

Variance(P::ewma)

This function returns the EWMA control variance.

source

ChemometricsTools.Variance — Method.

Variance(rv::RunningVar)

Returns the current variance inside of a RunningVar object.

source

ChemometricsTools.entropy — Method.

entropy(v)

Calculates the Shannon-Entropy of a probability vector v. Returns a scalar. A common gain function used in tree methods.

source

ChemometricsTools.findpeaks — Method.

findpeaks( vY; m = 3)

Finds the indices of peaks in a vector vY with a window span of 2m. Original R function by Stas_G:(https://stats.stackexchange.com/questions/22974/how-to-find-local-peaks-valleys-in-a-series-of-data) This version is based on a C++ variant by me.

source

ChemometricsTools.gini — Method.

gini(p)

Calculates the GINI coefficient of a probability vector p. Returns a scalar. A common gain function used in tree methods.

source

ChemometricsTools.offsetToZero — Method.

offsetToZero(X)

Ensures that no observation(row) of Array X is less than zero, by ensuring the minimum value of each row is zero.

source

ChemometricsTools.rbinomial — Method.

rbinomial( p, size... )

Makes an N-dimensional array of size(s) size with a probability of being a 1 over a 0 of 1 p.

source

ChemometricsTools.ssd — Method.

ssd(p)

Calculates the sum squared deviations from a decision tree split. Accepts a vector of values, and the mean of that vector. Returns a scalar. A common gain function used in tree methods.

source

ChemometricsTools.DirectStandardizationXform — Method.

(DSX::DirectStandardizationXform)(X; Factors = length(DSX.pca.Values))

Applies a the transform from a learned direct standardization object DSX to new data X.

source

ChemometricsTools.ewma — Method.

EWMA(P::ewma)(New; train = true)

Provides an EWMA score for a New scalar value. If train == true the model is updated to include this new value.

source

ChemometricsTools.ChangeCenter — Method.

ChangeCenter(P::ewma, new::Float64)

This is a convenience function to update the center of a P EWMA model, to a new scalar value.

source

ChemometricsTools.ScaleMinMax — Method.

ScaleMinMax(X)

Scales the columns of X by the Min and Max of each row such that no observation is greater than 1 or less than zero. Returns the scaled array.

source

ChemometricsTools.boxcar — Method.

boxcar(X; windowsize = 3, fn = mean)

Applies a boxcar function (fn) to each window of size windowsize to every row in X.

source