Distance Measures

Distances API Reference

Functions

ChemometricsTools.Kernel — Method.

(K::Kernel)(X)

This is a convenience function to allow for one-line construction of kernels from a Kernel object K and new data X.

source

ChemometricsTools.CauchyKernel — Method.

CauchyKernel(X, Y, sigma)

Creates a Cauchy kernel from Arrays X and Y using hyperparameters sigma.

source

ChemometricsTools.CauchyKernel — Method.

CauchyKernel(X, sigma)

Creates a Cauchy kernel from Array X using hyperparameters sigma.

source

ChemometricsTools.EuclideanDistance — Method.

EuclideanDistance(X, Y)

Returns the euclidean distance matrix of X and Y such that the columns are the samples in Y.

source

ChemometricsTools.EuclideanDistance — Method.

EuclideanDistance(X)

Returns the Grahm aka the euclidean distance matrix of X.

source

ChemometricsTools.GaussianKernel — Method.

GaussianKernel(X, Y, sigma)

Creates a Gaussian/RBF kernel from Arrays X and Y with hyperparameter sigma.

source

ChemometricsTools.GaussianKernel — Method.

GaussianKernel(X, sigma)

Creates a Gaussian/RBF kernel from Array X using hyperparameter sigma.

source

ChemometricsTools.LinearKernel — Method.

LinearKernel(X, Y, c)

Creates a Linear kernel from Arrays X and Y with hyperparameter C.

source

ChemometricsTools.LinearKernel — Method.

LinearKernel(X, c)

Creates a Linear kernel from Array X and hyperparameter C.

source

ChemometricsTools.ManhattanDistance — Method.

ManhattanDistance(X, Y)

Returns the Manhattan distance matrix of X and Y such that the columns are the samples in Y.

source

ChemometricsTools.ManhattanDistance — Method.

ManhattanDistance(X)

Returns the Manhattan distance matrix of X.

source

ChemometricsTools.SquareEuclideanDistance — Method.

SquareEuclideanDistance(X, Y)

Returns the squared euclidean distance matrix of X and Y such that the columns are the samples in Y.

source

ChemometricsTools.SquareEuclideanDistance — Method.

SquareEuclideanDistance(X)

Returns the squared Grahm aka the euclidean distance matrix of X.

source